refactor(flutter): remove AppBarButtonLogout and separate progress_widget
