+++ packages/flutter/lib/src/components/legacy/canvas.dart
@@ -92,20 +92,3 @@ class LoadingOverlay extends StatelessWidget {
     );
   }
 }
-
-class ProgressWidget extends StatelessWidget {
-  const ProgressWidget({
-    Key key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
-        child: CircularProgressIndicator(
-          valueColor: AlwaysStoppedAnimation<Color>(ColorPalette.baseGrey50),
-        ),
-      ),
-    );
-  }
-}
